More quaterion=>Euler work

+++ src/imus/usfs.hpp
@@ -100,9 +100,9 @@ namespace hf {
 
             void computeEulerAngles(float q0, float q1, float q2, float q3, float euler[3])
             {
-                euler[0] = atan2(2.0f*(q0*q1+q2*q3),q0*q0-q1*q1-q2*q2+q3*q3);   // roll
-                euler[1] =  asin(2.0f*(q1*q3-q0*q2));                           // pitch
-                euler[2] = atan2(2.0f*(q1*q2+q0*q3),q0*q0+q1*q1-q2*q2-q3*q3);   // yaw (heading)
+                euler[0] = atan2f(2.0f*(q0*q1 + q2*q3), q0*q0 - q1*q1 - q2*q2 + q3*q3);   // roll
+                euler[1] =  asinf(2.0f*(q1*q3 - q0*q2));                                  // pitch
+                euler[2] = atan2f(2.0f*(q1*q2 + q0*q3), q0*q0 + q1*q1 - q2*q2 - q3*q3);   // yaw (heading)
 	     }
 
      }; // class USFS
